Add Return value constraint class generation TOMEE-2517 Bean Validation with MicroProfile JWT
